Fabric.js обновление логики #7 разделение на компоненты Layers.vue и Tools.vue
